Добавляет лбработчик на Esc, разделяет модули
