+++ js/form-validation.js
@@ -1,11 +1,17 @@
-import { sendData, getSurverDataSuccess, getSurverDataFail } from './api.js';
+import { showErrorUploadMessage } from './error-messages.js';
 import { isEscapeKey } from './utile.js';
+import { showSuccessMessage } from './error-messages.js';
+import { getSliderValue } from './photo-effects.js';
+// import { closeUploadForm } from './open-submit-form.js';
+import { sendData } from './api.js';
+
 
 const HASHTAG_START = '#';
 const HASHTAG_MIN_LENGTH = 2;
 const HASHTAG_MAX_LENGTH = 19;
 
 const uploadFileForm = document.querySelector('.img-upload__form');
+const imageUploadPreview = uploadFileForm.querySelector('.img-upload__preview');
 const hashtagInput = uploadFileForm.querySelector('[name="hashtags"]');
 const commentInput = uploadFileForm.querySelector('[name="description"]');
 
@@ -18,16 +24,28 @@ const isUnique = (arr) => new Set(arr).size === arr.length;
 
 const serializeHashtags = (value) => value.trim().toLowerCase().split(/\s+/);
 
+const getSurverDataSuccess = (response) => {
+  if(response.ok) {
+    showSuccessMessage();
+    // closeUploadForm();
+    uploadFileForm.reset();
+    imageUploadPreview.style.filter = '';
+    imageUploadPreview.style.transform = 'scale(100%)';
+    getSliderValue();
+  }
+};
+
 const getFormSubmit = () => {
   uploadFileForm.addEventListener('submit', (evt) => {
     evt.preventDefault();
     const isValid = pristine.validate();
     if (isValid) {
       const formData = new FormData(evt.target);
-      sendData(formData, getSurverDataSuccess, getSurverDataFail);
+      sendData(formData, getSurverDataSuccess, showErrorUploadMessage);
     }
   });
 };
+
 
 const stopPropagationEsc = (evt) => {
   if(isEscapeKey(evt)) {
